fixes
added disco ball for 5 strike

+++ puffy puffs/Assets/Scripts/board.cs
@@ -8,7 +8,7 @@ public class Board : MonoBehaviour
     public int width;
     public int height;
 
-    public int borderSize;
+    public float borderSize;
 
     public GameObject tileNormalPrefab;
     public GameObject tileObstaclePrefab;
